WPD: enable diode + pulse-train cases, consolidate scripts, wire CI
Now that the SPICE diode conducts (#197) and PULSE repeats (#196), enable the
full work-precision suite: filter (analytic golden), rc pulse-train (analytic),
graetz rectifier (VACASK golden), mul multiplier (Cadnip IDA golden — VACASK's
multiplier golden aborts at small steps).

- Consolidate the three stage scripts into run_wpd.jl + wpd_common.jl.
- Report renders inline ASCII work-precision diagrams (UnicodePlots), so it
  shows directly in the CI job summary — no image hosting. Drop Plots/GR.
- Per-case golden pinned in config.json (analytic|vacask|cadnip), chosen by what
  converges; run_wpd errors if the pinned golden fails (no silent fallback).
- Error measured at each run's own timepoints vs a dense reference, so
  high-order solvers aren't penalised for interpolating their sparse output.
- VACASK run at high-order Gear/BDF (tran_maxord=5), not its 2nd-order default.
- Split VACASK download into fetch_vacask.sh (shared by run_vacask.sh and CI).
- CI: new work-precision job (Julia 1.12); drop 1.11 from integration/benchmark
  jobs, keep it only in test-core for compat.
