+++ .github/workflows/ci.yml
@@ -70,7 +70,6 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.11'
           - '1.12'
 
     steps:
