+++ benchmarks/Project.toml
@@ -9,11 +9,11 @@ OrdinaryDiffEqBDF = "6ad6398a-0878-4a85-9266-38940aa047c8"
 OrdinaryDiffEqRosenbrock = "43230ef6-c299-4910-a778-202eb28ce4ce"
 OrdinaryDiffEqSDIRK = "2d112036-d095-4a1e-ab9a-08536f3ecdbf"
 PSPModels = "abf2981e-f194-4833-b9e6-388029af7f41"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Sundials = "c3572dad-4567-51f8-b174-8c6c989267f4"
+UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 VACASKModels = "a2b3c4d5-6789-4abc-def0-123456789abc"
 VADistillerModels = "2129788e-3e5d-4aba-9c60-0ef5ee775297"
