+++ benchmarks/vacask/wpd/.gitignore
@@ -1,5 +1,2 @@
-# Generated by cadnip_wpd.jl / vacask_wpd.jl / plot_wpd.jl
+# Generated by run_wpd.jl
 out/
-plots/
-__pycache__/
-*.pyc
