+++ benchmarks/vacask/wpd/config.json
@@ -1,47 +1,46 @@
 {
-  "_comment": "Shared work-precision-diagram config, read by cadnip_wpd.jl, vacask_wpd.jl and plot_wpd.jl. Keep grid/tspan/sweep identical so Cadnip and VACASK are compared on the same time grid against the same golden reference.",
+  "_comment": "Shared work-precision-diagram config, read by wpd_common.jl / run_wpd.jl. Keep tspan/sweep identical across simulators so error is compared consistently against the same golden.",
 
   "reltols": [1e-3, 1e-4, 1e-5, 1e-6, 1e-7, 1e-8, 1e-9],
   "abstol_scale": 1.0,
 
-  "_ref_comment": "Golden = VACASK at ref_reltol with a fine maxstep (tspan/(n_grid*ref_maxstep_factor)) so the reference trajectory is accurate without tripping VACASK's min-timestep abort, which fires at very tight reltol on stiff diode switching. ref_abstol is used as VACASK's vntol.",
+  "_ref_comment": "The VACASK golden uses ref_reltol with a fine maxstep (tspan/(n_grid*ref_maxstep_factor)) so the trajectory is accurate without tripping VACASK's min-timestep abort. ref_abstol is used as VACASK's vntol.",
   "ref_reltol": 1e-7,
   "ref_abstol": 1e-7,
   "ref_maxstep_factor": 50,
 
-  "_vacask_integration_comment": "VACASK defaults to trapezoidal (2nd order). Use its variable-order Gear/BDF (up to 5th order) so VACASK is benchmarked at its best, not artificially limited to 2nd order.",
+  "_vacask_integration_comment": "VACASK defaults to trapezoidal (2nd order). Use its variable-order Gear/BDF (up to 5th order) so it is benchmarked at its best.",
   "vacask_tran_method": "gear",
   "vacask_tran_maxord": 5,
 
   "n_grid": 2000,
 
-  "_rc_comment": "rc tspan is one pulse period only (pulse rises at t=1us, high for 1ms). Cadnip's PULSE source does not repeat (the 'period' parameter is ignored - it fires once then stays at val0), so the window is limited to the first pulse where Cadnip and VACASK agree. This still exercises adaptive discontinuity handling at the two sharp edges. Widen once the Cadnip pulse-repetition bug is fixed.",
+  "_golden_comment": "golden pins the reference per case (analytic|vacask|cadnip), chosen by what actually converges: filter/rc have exact closed forms; graetz's tight VACASK run completes; mul's VACASK golden aborts (min-timestep) so it uses a tight Cadnip IDA reference. run_wpd.jl errors if the pinned golden fails - it never silently substitutes.",
 
   "cases": {
     "filter": {
-      "title": "3rd-order Butterworth LC filter (linear, analytic)",
+      "title": "3rd-order Butterworth LC filter (linear)",
       "tspan": [0.0, 100.0],
       "output": ["vout"],
-      "linear": true,
-      "analytic": true
+      "golden": "analytic"
     },
     "rc": {
-      "title": "RC circuit, single pulse step response (linear)",
-      "tspan": [0.0, 0.002],
+      "title": "RC circuit driven by a pulse train (linear)",
+      "tspan": [0.0, 0.02],
       "output": ["2"],
-      "linear": true,
-      "analytic": false
-    }
-  },
-
-  "_disabled_cases_comment": "graetz (and the diode voltage multiplier) are wired up end-to-end but DISABLED because Cadnip's SPICE diode model currently does not conduct (a single diode reads as an open circuit in DC; the rectified output is 0). VACASK simulates them correctly. Re-enable by moving an entry back into 'cases' once the Cadnip diode model is fixed. The voltage multiplier additionally trips VACASK's min-timestep abort below reltol~1e-5, so it is a poor work-precision case regardless.",
-  "_disabled_cases": {
+      "golden": "analytic"
+    },
     "graetz": {
-      "title": "Graetz bridge full-wave rectifier (nonlinear)",
+      "title": "Graetz bridge full-wave rectifier (nonlinear diode)",
       "tspan": [0.0, 0.1],
       "output": ["outp", "outn"],
-      "linear": false,
-      "analytic": false
+      "golden": "vacask"
+    },
+    "mul": {
+      "title": "Diode voltage multiplier (nonlinear, stiff)",
+      "tspan": [0.0, 5e-4],
+      "output": ["20"],
+      "golden": "cadnip"
     }
   }
 }
